bibsonomy-rest-common/src/main/java/org/bibsonomy/rest/renderer/RendererFactory.java
/**
 *  
 *  BibSonomy-Rest-Common - Common things for the REST-client and server.
 *   
 *  Copyright (C) 2006 - 2010 Knowledge & Data Engineering Group, 
 *                            University of Kassel, Germany
 *                            http://www.kde.cs.uni-kassel.de/
 *  
 *  This program is free software; you can redistribute it and/or
 *  modify it under the terms of the GNU Lesser General Public License
 *  as published by the Free Software Foundation; either version 2
 *  of the License, or (at your option) any later version.
 * 
 *  This program is distributed in the hope that it will be useful,
 *  but WITHOUT ANY WARRANTY; without even the implied warranty of
 *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *  GNU Lesser General Public License for more details.
 *  
 *  You should have received a copy of the GNU Lesser General Public License
 *  along with this program; if not, write to the Free Software
 *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */

package org.bibsonomy.rest.renderer;

import org.bibsonomy.common.exceptions.InternServerException;
import org.bibsonomy.rest.enums.RenderingFormat;
import org.bibsonomy.rest.renderer.impl.JSONRenderer;
import org.bibsonomy.rest.renderer.impl.XMLRenderer;

/**
 * A factory to get implementations of the
 * {@link org.bibsonomy.rest.renderer.Renderer}-interface.
 * 
 * @author Christian Schenk
 * @version $Id$
 */
public class RendererFactory {
	
	private static Renderer JSON_RENDERER = new JSONRenderer();
	private static Renderer XML_RENDERER = new XMLRenderer();

	/**
	 * Returns the renderer for the given format; it defaults to the XML renderer.
	 * @param renderingFormat 
	 * @return the renderer
	 */
	public static Renderer getRenderer(final RenderingFormat renderingFormat) {
		if (renderingFormat == null) throw new InternServerException("RenderingFormat is null");
		
		if (RenderingFormat.JSON.equals(renderingFormat)) {
			return JSON_RENDERER;
		}
		
		// default is xml renderer
		return XML_RENDERER;
	}
}